Various fixes to ensure system-probe tests cleanup properly (#17120)
* Cleanup goroutines from process cache and log limits

* Fix Conntrack test goroutine leaks

* Fix DNS cache close

* Add missing TCP server shutdown

* Stop process monitor in TestProcessMonitorInNamespace

* Fix dns statkeeper close

* Fix circuit break ticker

* Ensure mysql db connection always closed

* Use non-hardcoded netns names in TestGatewayLookupCrossNamespace

* Use standard DNAT setup helper

* Ensure http server is closed

* Increase kafka start timeout to 2 minutes

* close output in compile tests

* Ensure cleanup in serverutils

* Ensure mysql has the right type

* Fix iptables restore usage

* Ensure conntracker is closed

* Ensure correct kafka type

* empty commit

+++ pkg/network/usm/kafka_monitor_test.go
@@ -103,11 +103,12 @@ func TestKafkaProtocolParsing(t *testing.T) {
 		if _, ok := ctx.extras["client"]; !ok {
 			return
 		}
-		client := ctx.extras["client"].(*kafka.Client)
-		defer client.Client.Close()
-		for k, value := range ctx.extras {
-			if strings.HasPrefix(k, "topic_name") {
-				_ = client.DeleteTopic(value.(string))
+		if client, ok := ctx.extras["client"].(*kafka.Client); ok {
+			defer client.Client.Close()
+			for k, value := range ctx.extras {
+				if strings.HasPrefix(k, "topic_name") {
+					_ = client.DeleteTopic(value.(string))
+				}
 			}
 		}
 	}
@@ -290,6 +291,7 @@ func TestKafkaProtocolParsing(t *testing.T) {
 
 				serverAddr := "localhost:8081"
 				srvDoneFn := testutil.HTTPServer(t, "localhost:8081", testutil.Options{})
+				t.Cleanup(srvDoneFn)
 				httpClient := nethttp.Client{}
 
 				req, err := nethttp.NewRequest(httpMethods[0], fmt.Sprintf("http://%s/%d/request", serverAddr, nethttp.StatusOK), nil)
